Inline installer GC cleanup helper
