+++ pkg/fleet/installer/exec/BUILD.bazel
@@ -20,7 +20,6 @@ go_library(
     ] + select({
         "@rules_go//go/platform:windows": [
             "//pkg/fleet/installer/db",
-            "//pkg/fleet/installer/garbagecollect",
         ],
         "//conditions:default": [],
     }),
