--- pkg/fleet/installer/garbagecollect/BUILD.bazel
@@ -1,23 +1,0 @@
-load("@rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "garbagecollect",
-    srcs = ["garbagecollect.go"],
-    importpath = "github.com/DataDog/datadog-agent/pkg/fleet/installer/garbagecollect",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/fleet/installer/repository",
-        "//pkg/util/log",
-    ],
-)
-
-go_test(
-    name = "garbagecollect_test",
-    srcs = ["garbagecollect_test.go"],
-    embed = [":garbagecollect"],
-    gotags = ["test"],
-    deps = [
-        "@com_github_stretchr_testify//assert",
-        "@com_github_stretchr_testify//require",
-    ],
-)
