+++ pkg/fleet/installer/repository/repository_test.go
@@ -11,6 +11,7 @@ import (
 	"path"
 	"path/filepath"
 	"testing"
+	"time"
 
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/symlink"
 	"github.com/stretchr/testify/assert"
@@ -404,6 +405,50 @@ func TestRepairDirectoryDifferentCasing(t *testing.T) {
 
 	err := repairDirectory(sourceDir, targetDir)
 	assert.Error(t, err)
+}
+
+func TestCleanupTmpDirectory(t *testing.T) {
+	tmpDir := t.TempDir()
+	entries := []struct {
+		name       string
+		path       string
+		isDir      bool
+		age        time.Duration
+		wantExists bool
+	}{
+		{name: "old file", path: filepath.Join(tmpDir, "old-file"), age: -25 * time.Hour},
+		{name: "old directory", path: filepath.Join(tmpDir, "old-dir"), isDir: true, age: -25 * time.Hour},
+		{name: "recent file", path: filepath.Join(tmpDir, "recent-file"), age: -1 * time.Hour, wantExists: true},
+	}
+
+	for _, entry := range entries {
+		if entry.isDir {
+			assert.NoError(t, os.MkdirAll(filepath.Join(entry.path, "subdir"), 0755))
+		} else {
+			assert.NoError(t, os.WriteFile(entry.path, []byte(entry.name), 0644))
+		}
+		entryTime := time.Now().Add(entry.age)
+		assert.NoError(t, os.Chtimes(entry.path, entryTime, entryTime))
+	}
+
+	assert.NoError(t, CleanupTmpDirectory(tmpDir))
+
+	for _, entry := range entries {
+		entry := entry
+		t.Run(entry.name, func(t *testing.T) {
+			if entry.wantExists {
+				assert.FileExists(t, entry.path)
+			} else if entry.isDir {
+				assert.NoDirExists(t, entry.path)
+			} else {
+				assert.NoFileExists(t, entry.path)
+			}
+		})
+	}
+}
+
+func TestCleanupTmpDirectoryIgnoresMissingDirectory(t *testing.T) {
+	assert.NoError(t, CleanupTmpDirectory(filepath.Join(t.TempDir(), "missing")))
 }
 
 // This test is used to verify that the repository can handle external packages that are symlinked.
